feat: restructure screens, tables and modals
