+++ src/domain/publishable-api-keys/modals/manage-sales-channels.tsx
@@ -1,10 +1,4 @@
 import React, { useEffect, useState } from "react"
-import {
-  TableOptions,
-  usePagination,
-  useRowSelect,
-  useTable,
-} from "react-table"
 import { AnimatePresence, motion } from "framer-motion"
 
 import { SalesChannel } from "@medusajs/medusa"
@@ -20,196 +14,15 @@ import SideModal from "../../../components/molecules/modal/side-modal"
 import CrossIcon from "../../../components/fundamentals/icons/cross-icon"
 import useNotification from "../../../hooks/use-notification"
 import InputField from "../../../components/molecules/input"
-import IndeterminateCheckbox from "../../../components/molecules/indeterminate-checkbox"
-import TableContainer from "../../../components/organisms/table-container"
-import Table from "../../../components/molecules/table"
 import SearchIcon from "../../../components/fundamentals/icons/search-icon"
-import { TablePagination } from "../../../components/organisms/table-container/pagination"
 import BackIcon from "../../../components/fundamentals/icons/back-icon"
-
-/* ************************************* */
-/* *************** TABLE *************** */
-/* ************************************* */
+import SalesChannelTable from "../tables/sales-channels-table"
 
 const LIMIT = 12
 
-const COLUMNS = [
-  {
-    width: 30,
-    id: "selection",
-    Header: ({ getToggleAllPageRowsSelectedProps }) => (
-      <span className="flex justify-center w-[30px]">
-        <IndeterminateCheckbox {...getToggleAllPageRowsSelectedProps()} />
-      </span>
-    ),
-    Cell: ({ row }) => {
-      return (
-        <span
-          onClick={(e) => e.stopPropagation()}
-          className="flex justify-center w-[30px]"
-        >
-          <IndeterminateCheckbox {...row.getToggleRowSelectedProps()} />
-        </span>
-      )
-    },
-  },
-  {
-    Header: "Title",
-    accessor: "name",
-  },
-  {
-    Header: "Description",
-    accessor: "description",
-  },
-]
-
-type SalesChannelTableProps = {
-  isEdit: boolean
-  query: string
-  selectedChannels: Map<string, SalesChannel>
-  setSelectedChannels: (
-    setter: (oldState: Map<string, SalesChannel>) => void
-  ) => {}
-  offset: number
-  setOffset: (offset: number) => void
-  data: SalesChannel[]
-  isLoading: boolean
-  count: number
-}
-
-/**
- * Sales channels select table.
- */
-function SalesChannelTable(props: SalesChannelTableProps) {
-  const {
-    isEdit,
-    query,
-    offset,
-    data,
-    isLoading,
-    count,
-    setOffset,
-    selectedChannels,
-    setSelectedChannels,
-  } = props
-
-  const tableConfig: TableOptions<SalesChannel> = {
-    data,
-    // @ts-ignore
-    columns: COLUMNS,
-    autoResetPage: false,
-    manualPagination: true,
-    autoResetSelectedRows: false,
-    initialState: {
-      pageIndex: Math.floor(offset / LIMIT),
-      pageSize: LIMIT,
-      selectedRowIds: Object.keys(selectedChannels).reduce((prev, k) => {
-        prev[k] = true
-        return prev
-      }, {}),
-    },
-    pageCount: Math.ceil(count / LIMIT),
-    getRowId: (row) => row.id,
-  }
-
-  const table = useTable(tableConfig, usePagination, useRowSelect)
-
-  useEffect(() => {
-    setSelectedChannels((oldState) => {
-      const newState = {}
-
-      Object.keys(table.state.selectedRowIds).forEach((k) => {
-        newState[k] = oldState[k] || data.find((d) => d.id === k)
-      })
-      return newState
-    })
-  }, [table.state.selectedRowIds, data])
-
-  useEffect(() => {
-    setOffset(0)
-    table.gotoPage(0)
-  }, [query])
-
-  const handleNext = () => {
-    if (table.canNextPage) {
-      setOffset(offset + LIMIT)
-      table.nextPage()
-    }
-  }
-
-  const handlePrev = () => {
-    if (table.canPreviousPage) {
-      setOffset(Math.max(offset - LIMIT, 0))
-      table.previousPage()
-    }
-  }
-
-  return (
-    <>
-      <TableContainer hasPagination numberOfRows={LIMIT} isLoading={isLoading}>
-        <Table {...table.getTableProps()}>
-          <Table.Head>
-            {table.headerGroups?.map((headerGroup) => (
-              <Table.HeadRow {...headerGroup.getHeaderGroupProps()}>
-                {headerGroup.headers.map((col) => (
-                  <Table.HeadCell {...col.getHeaderProps()}>
-                    {col.render("Header")}
-                  </Table.HeadCell>
-                ))}
-              </Table.HeadRow>
-            ))}
-          </Table.Head>
-          <Table.Body {...table.getTableBodyProps()}>
-            {table.rows.map((row) => {
-              table.prepareRow(row)
-              return (
-                <Table.Row color={"inherit"} {...row.getRowProps()}>
-                  {row.cells.map((cell) => {
-                    return (
-                      <Table.Cell {...cell.getCellProps()}>
-                        {cell.render("Cell")}
-                      </Table.Cell>
-                    )
-                  })}
-                </Table.Row>
-              )
-            })}
-          </Table.Body>
-        </Table>
-      </TableContainer>
-      <div className="absolute w-[506px]" style={{ bottom: 100 }}>
-        <TablePagination
-          pagingState={{
-            count,
-            offset,
-            title: "Sales Channels",
-            pageSize: offset + table.rows.length,
-            currentPage: table.state.pageIndex + 1,
-            pageCount: table.pageCount,
-            nextPage: handleNext,
-            prevPage: handlePrev,
-            hasNext: table.canNextPage,
-            hasPrev: table.canPreviousPage,
-          }}
-          isLoading={isLoading}
-        />
-      </div>
-    </>
-  )
-}
-
 /* ****************************************** */
-/* *************** SIDE MODAL *************** */
+/* *************** ADD SCREEN *************** */
 /* ****************************************** */
-
-type ManageSalesChannelsSideModalProps = {
-  keyId?: string
-  close: () => void
-  isVisible: boolean
-  setSelectedChannels: (arg: any) => void
-  selectedChannels: Map<string, SalesChannel>
-  isEdit?: boolean
-}
 
 function AddScreen(props: {
   isEdit?: boolean
@@ -329,6 +142,13 @@ function AddScreen(props: {
   )
 }
 
+/* ******************************************* */
+/* *************** EDIT SCREEN *************** */
+/* ******************************************* */
+
+/**
+ * Edit exiting PK SC list
+ */
 function EditScreen(props: {
   isEdit?: boolean
   keyId: string
@@ -460,8 +280,21 @@ function EditScreen(props: {
   )
 }
 
+/* ***************************************************** */
+/* *************** MANAGE CHANNELS MODAL *************** */
+/* ***************************************************** */
+
+type ManageSalesChannelsSideModalProps = {
+  keyId?: string
+  close: () => void
+  isVisible: boolean
+  setSelectedChannels: (arg: any) => void
+  selectedChannels: Map<string, SalesChannel>
+  isEdit?: boolean
+}
+
 /**
- * Publishable Key details container.
+ * Modal for adding/removing existing PKs channels.
  */
 function ManageSalesChannelsSideModal(
   props: ManageSalesChannelsSideModalProps
